+++ src/utils/api.js
@@ -1,6 +1,6 @@
 import axios from 'axios';
 
-import { SERVER_URL } from 'constants';
+import { SERVER_URL } from 'constants/constants';
 
 export const getProductList = async () => {
   try {
